Move path helpers into a module to be shared by all controller specs
