+++ spec/controllers/api/v0x0/sources_controller_spec.rb
@@ -48,24 +48,4 @@ RSpec.describe Api::V0x0::SourcesController, type: :controller do
     expect(response.location).to match(a_string_ending_with("api/v0.0/sources/#{source.id}"))
     expect(JSON.parse(response.parsed_body)).to include("name" => "abc", "id" => source.id.to_s)
   end
-
-  def delete_path(path)
-    parsed_params = Rails.application.routes.recognize_path(path, :method => "DELETE")
-    delete(parsed_params[:action], :params => parsed_params.except(:action, :controller))
-  end
-
-  def get_path(path)
-    parsed_params = Rails.application.routes.recognize_path(path)
-    get(parsed_params[:action], :params => parsed_params.except(:action, :controller))
-  end
-
-  def patch_path(path, params)
-    parsed_params = Rails.application.routes.recognize_path(path, :method => "PATCH")
-    patch(parsed_params[:action], :params => params[:params].merge(parsed_params.except(:action, :controller)))
-  end
-
-  def post_path(path, options)
-    parsed_params = Rails.application.routes.recognize_path(path, :method => "POST")
-    post(parsed_params[:action], options)
-  end
 end
