+++ spec/spec_helper.rb
@@ -1,3 +1,5 @@
+Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
+
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
@@ -93,4 +95,5 @@ RSpec.configure do |config|
   # as the one that triggered the failure.
   Kernel.srand config.seed
 =end
+  config.include Spec::Support::PathHelpers, :type => :controller
 end
